requirements.txt
pynput==1.4
pywin32==224
opencv_python==4.1.0.25
numpy==1.16.3
pytesseract==0.2.6
waiting==1.4.1
matplotlib==3.0.3
Pillow==7.0.0
win32gui==221.6
